Retired 'Relay worker comment notifications to email' guide
The guide "Include survey links on outgoing messages with bots" is no longer relevant as it has been replaced with a workflow. An edited version of the old guide's intro has been added to the new workflow doc.

The guide and its assets have been removed and a redirect set for ~~~workflows/cerb.notifications.mention_emailer/~~~.

Fixes #89